requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
curl_cffi>=0.5.0
idna==3.10
loguru==0.7.3
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.3.0
